Align vendored code closer to the source
