+++ cmd/src/code_intel_upload.go
@@ -84,7 +84,7 @@ func handleCodeIntelUpload(args []string) error {
 	uploadOptions := codeintelUploadOptions(out)
 	var uploadID int
 	if codeintelUploadFlags.gzipCompressed {
-		uploadID, err = UploadCompressedIndex(ctx, codeintelUploadFlags.file, client, uploadOptions)
+		uploadID, err = UploadCompressedIndex(ctx, codeintelUploadFlags.file, client, uploadOptions, 0)
 	} else {
 		uploadID, err = UploadUncompressedIndex(ctx, codeintelUploadFlags.file, client, uploadOptions)
 	}
